feat: 구조 추가 및 SwiftUI ChartView 수정
